[uart_drop] rename for uniformity: module name='switch' type='uart'
