+++ sw/airborne/modules/switching/auto1_commands.c
