+++ sw/airborne/modules/switching/switch_uart.c
@@ -18,17 +18,35 @@
  * <http://www.gnu.org/licenses/>.
  */
 /**
- * @file "modules/com/uart_drop.h"
+ * @file "modules/com/uart_drop.c"
  * @author Freek van Tienen
  * Module for dropping balls using UART
  */
 
-#include "std.h"
+#include "modules/switching/switch_uart.h"
+#include "mcu_periph/uart.h"
 
-#ifndef UART_DROP_H
-#define UART_DROP_H
+uint8_t switch_uart_channel = 0;
 
-extern void drop_ball(uint8_t number);
+static uint8_t drop_string[] = "<Drop_Paintball_Now";
+#define DROP_STRINGLEN 19
 
-#endif
+void drop_ball(uint8_t number) {
+  for(uint8_t i = 0; i < DROP_STRINGLEN; i++)
+    uart_put_byte(&SWITCH_UART_PORT, 0, drop_string[i]);
 
+  // Drop next ball
+  uint8_t last = '>';
+
+  // Or Drop a specific ball
+  if(number == 1) {
+    last = '1';
+  } else if(number == 2) {
+    last = '2';
+  } else if(number == 3) {
+    last = '3';
+  } else if(number == 4) {
+    last = '4';
+  }
+  uart_put_byte(&SWITCH_UART_PORT, 0, last);
+}
